Update input component

+++ src/components/core/controls/input/index.js
@@ -5,6 +5,14 @@ import styles from './styles.scss';
 
 const Input = (props) => {
   const { label, labelClassname, containerClassname, truncateLabel, name,  ...inputProps } = props;
+  if (!label) {
+    return (
+      <input
+        name={name}
+        {...inputProps}
+      />
+    );
+  }
   return (
     <label className={cx(styles.container, containerClassname)}>
       <span className={cx(styles.label, truncateLabel && 'text-truncate ', labelClassname)}>{label}</span>
